requirements.txt
SQLAlchemy==1.3.23
python-telegram-bot==13.4.1